[UT] more coverage LedgerMongodb 68.44% Ref #80

+++ src/networking/tcp/HeaderPattern.hpp
@@ -6,7 +6,7 @@ namespace networking {
 namespace tcp {
 
 struct __attribute__((__packed__)) HeaderPattern {
-  uint16_t size;
+  uint32_t size;
   ProtocolType type;
   uint8_t signature[64];
 };
